requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
sktime
holidays
convertdate
psycopg2-binary
sqlalchemy
python-dotenv
pydantic